requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
deepdiff==7.0.1
exceptiongroup==1.2.2
fastapi==0.112.1
frozendict==2.4.4
frozenlist==1.4.1
h11==0.14.0
idna==3.7
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
multidict==6.0.5
ordered-set==4.1.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
PyLD==2.0.4
python-dotenv==1.0.1
referencing==0.35.1
requests==2.32.3
rfc3986-validator==0.1.1
rpds-py==0.20.0
ruff==0.6.1
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
validators==0.33.0
yarl==1.9.4
